+++ gateway/README.md
@@ -1,3 +1,7 @@
 # BUGOUT gateway
 
 websocket server which negotiates between BUGOUT frontend and an internal kafka cluster serving the [judge](../judge/README.md).
+
+## Attribution
+
+Thank you to the authors of [ws-rs](https://github.com/housleyjk/ws-rs).  We adapted your examples to fit this use case and appreciate the effort in releasing this library.
